feat: 기회인지 생성 (연관 테이블 아직 안함)

+++ src/main/java/be15fintomatokatchupbe/campaign/command/application/dto/request/CreateChanceRequest.java
@@ -18,13 +18,13 @@ public class CreateChanceRequest {
     private Long clientManagerId;
     private LocalDateTime startedAt;
     private LocalDateTime  endedAt;
-    private List<Integer> userList;
+    private List<Long> userList;
     private String awarenessPath;
     private String productName;
     private Long productPrice;
     private Long expectedRevenue;
     private BigDecimal expectedProfitMargin;
     private String notes;
-    private List<Integer> categoryList;
+    private List<Long> categoryList;
 }
